server/python/requirements.txt
flask
numpy
pillow
torch
torchvision
matplotlib
segmentation-models-pytorch
